requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.5.0
openai==1.46.0
Pillow==10.4.0
numpy==2.1.1
httpx==0.27.2
